Add At Home API Layer

+++ src/AtHome.Service/models/CompanyShippingDeal.cs
@@ -5,7 +5,7 @@ namespace AtHome.Service.models
     public class CompanyShippingDeal
     {
         public ShippingCompany Company {get; set;}
-        public double Amonut {get; set;} 
+        public decimal Amonut {get; set;} 
         
     }
 }
