+++ src/AtHome.ShippingCompanyAPI/handlers/ShippingCompanyAPI.cs
@@ -5,10 +5,10 @@ namespace AtHome.APIHandler.handlers
 {
     public class ShippingCompanyAPI : IShippingCompanyAPI
     {
-        public double GetOffer(ShippingCompany company, ShippingInfo shipingInfo)
+        public decimal GetOffer(ShippingCompany company, ShippingInfo shipingInfo)
         {
             Random random =new Random();
-            return random.NextDouble() * (100 - 0) + 0;
+            return (decimal)random.NextDouble() * (100 - 0) + 0;
         }
     }
 }
